Refactor database code out of HTTP handlers
This commit removes all database specific code out of the HTTP handlers.
It defines a new interface Repo that will allow us to support additional
databases or backends as necessary. Some additional changes will be
required for sure (such as adding ways to manage roles to the UserRepo)
but this code is stable and allows us to support the existing
functionality as is so is ready for master.