Complete App Layout
The Sidebar is now available as a general component and it's state (width,
whether or not to show it) are controlled via the store. You can include
and slot in whatever elements you would like inside of the sidebar.

Additionally all styling is now in line with what we had before (except with
the better vue bootstrap styles).